requirements.txt
# pip install pip==9.0.3
allennlp==0.9.0
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
scikit-learn==0.21.3
scispacy==0.2.2
simstring-pure==1.0.0
spacy==2.1.7
tensorboard==1.14.0
tensorboardX==1.8
tensorflow==1.14.0
tensorflow-estimator==1.14.0
torch==1.2.0
torchcontrib==0.0.2
torchvision==0.4.0
transformers==2.0.0